src/test/java/io/github/davemeier82/homeautomation/weewx/WeewxMessageTest.java
/*
 * Copyright 2021-2024 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *       https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package io.github.davemeier82.homeautomation.weewx;

import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.Test;

import java.io.IOException;

import static java.nio.charset.StandardCharsets.UTF_8;
import static org.apache.commons.io.IOUtils.resourceToString;
import static org.assertj.core.api.Assertions.assertThat;

class WeewxMessageTest {

  @Test
  void jsonDeserializationTest() throws IOException {
    WeewxMessage message = new ObjectMapper().readValue(resourceToString("/loop-payload.json", UTF_8), WeewxMessage.class);
    assertThat(message.getUv()).isEqualTo("1.0");
    assertThat(message.getLuminosityLux()).isEqualTo("19194.0");
  }

}